\name{MDES-package}
\alias{MDES-package}
\alias{MDES}
\docType{package}
\title{
Calculates the Minimum Detectable Effect Size (MDES)
~~ MDES ~~
}
\description{
This package calculates MDES for random assignment of individuals, with or without clustering, and for Quasi-Experimental Designs
}
\details{
\tabular{ll}{
Package: \tab MDES\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-11\cr
License: \tab ?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Ignacio Martinez

Maintainer: Ignacio Martinez <ignacio82@gmail.com>
~~ Ignacio Martinez ~~
}