% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDES_RA.R
\name{MDES_RA}
\alias{MDES_RA}
\title{MDES Random Assignment}
\usage{
MDES_RA(sig.lvl = 0.05, power = 0.8, N, p = 0.5, outcome = c("binary",
  "continuous"), mean_y, sigma_y, R_sq_xy = 0)
}
\arguments{
\item{sig.lvl}{Level of significance (default 0.05).}

\item{power}{Power (default 0.80).}

\item{N}{Total number of individuals in the sample.}

\item{p}{Probability of assignment to the treatment group (default 0.5).}

\item{outcome}{\code{"binary"} or \code{"continuous"}.}

\item{mean_y}{Mean of the outcome variable (required for binary outcomes)}

\item{sigma_y}{standard deviation of the outcome (required for continious outcomes).}

\item{R_sq_xy}{The proportion of the variance of outcome y explained by covariates x (default 0).}
}
\value{
A dataframe with \code{MDES} and \code{MDE}.
}
\examples{
MDES_RA(N=5000, outcome = "continuous", sigma_y = 0.1)
MDES_RA(N=5000, outcome = "binary", mean_y = 0.5)
}

