% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\title{MDES and MDE Calculator}
\usage{
power(type = c("Random Assigment", "MCG", "DID"), clusters = c(TRUE, FALSE),
  outcome = c("binary", "continuous"), sig.lvl = 0.05, power = 0.8, N, g,
  p = 0.5, mean_y, sigma_y, rho = 0.02, R_sq_WG = 0, R_sq_BG = 0,
  R_sq_xy = 0, R_sq_xyt = 0, R_sq_xt = 0.15, theta)
}
\arguments{
\item{type}{\code{c("Random Assigment", "MCG", "DID")}.}

\item{clusters}{\code{c(TRUE, FALSE)}}

\item{outcome}{\code{c("binary", "continuous")}}

\item{sig.lvl}{Level of significance (default 0.05).}

\item{power}{Power (default 0.80).}

\item{N}{Total number of individuals in the sample.}

\item{g}{Number of clusters.}

\item{p}{Probability of assignment to the treatment group (default 0.5).}

\item{mean_y}{Mean of the outcome variable (required for binary outcomes)}

\item{sigma_y}{standard deviation of the outcome (required for continious outcomes).}

\item{rho}{Intraclass correlation coefficient (default 0.02).}

\item{R_sq_WG}{Proportion of the within-group variance of outcome y explained by covariates x (default 0).}

\item{R_sq_BG}{Proportion of the group-level variance of outcome y explained by covariates x (default 0).}

\item{R_sq_xy}{The proportion of the variance of outcome y explained by covariates x (default 0).}

\item{R_sq_xyt}{Proportion of the individual-level variance of outcome y explained by covariates x (default 0).}

\item{R_sq_xt}{Proportion of the variance of the treatment indicator explained by covariates x (default 0.15).}

\item{theta}{The proportion of the outcome y variance due to persistent factors (0 <= theta <= 1). If the pre- and pos-observations are independent, then theta = 0.}
}
\value{
A dataframe with \code{MDES} and \code{MDE}.
}
\examples{
power(type = 'Random Assigment', clusters = FALSE, outcome = 'continuous', 
sig.lvl = .05, power = .8, N = 5000, p = 0.5, sigma_y = 0.1, R_sq_xy = 0)

power(type = 'Random Assigment', clusters = TRUE, outcome = 'continuous',
sig.lvl = .05, power = .8, N = 5000, g = 20, p = .5,
sigma_y = 0.1, rho = 0.02, R_sq_WG = 0, R_sq_BG = 0)

power(type = 'MCG', clusters = FALSE, outcome = 'continuous', sig.lvl = .05, 
power = .8, N = 5000, p = .5, sigma_y = 0.1, rho = 0.02, R_sq_xyt = 0, 
R_sq_xt = 0)

power(type = 'MCG', clusters = TRUE, outcome = 'continuous', sig.lvl = .05,
power = .8, N = 5000, g = 20, p = .5, sigma_y = 0.1, rho = 0.02,
R_sq_WG = 0, R_sq_xt = 0, R_sq_BG = 0)

power(type = 'DID', clusters = FALSE, outcome = 'continuous', sig.lvl = .05,
power = .8, N = 5000, p = .5, sigma_y = 0.1, rho = 0.02, R_sq_WG = 0, 
R_sq_xt = 0, theta = 0)

power(type = 'DID', clusters = TRUE, outcome = 'continuous', sig.lvl = .05, 
power = .8, N = 5000, g = 20, p = .5, sigma_y = 0.1, rho = 0.02, 
R_sq_WG = 0, R_sq_xt = 0, theta = 0, R_sq_BG = 0)
}

